requirements.txt
PyQt5
PyQt5-tools
alembic
sqlalchemy
mysql-connector-python